requirement.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
elastic-transport==8.15.1
elasticsearch==8.16.0
Faker==33.1.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1